part3/requirements.txt
Flask==3.0.0
pandas
numpy
topsis-himani-102303648==1.0.2
